# feat(es-dev-server): improve node-resolve performance