--- a/packages/es-dev-server/demo/performance/module-a.js
+++ b/packages/es-dev-server/demo/performance/module-a.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+import 'lodash-es';