--- a/packages/es-dev-server/test/fixtures/simple/src/symlinked-file.js
+++ b/packages/es-dev-server/test/fixtures/simple/src/symlinked-file.js
@@ -1,0 +1,1 @@
+../symlinked-dir/foo.js