… packages/building-utils/forked-babel-plugin-bare-import-rewrite.js
… packages/es-dev-server/README.md
… packages/es-dev-server/demo/performance/index.html
… packages/es-dev-server/package.json
… packages/es-dev-server/src/middleware/babel.js
… packages/es-dev-server/src/middleware/reload-browser.js
… packages/es-dev-server/src/utils/babel-compiler.js
… packages/es-dev-server/src/utils/caching-compiler.js
… packages/es-dev-server/src/utils/resolve-module-imports.js
… packages/es-dev-server/src/utils/setup-browser-reload.js
… packages/es-dev-server/src/utils/utils.js
… packages/es-dev-server/test/middleware/cache.test.js
… packages/es-dev-server/test/middleware/compile.test.js
… packages/es-dev-server/test/middleware/message-channel.test.js
… packages/es-dev-server/test/middleware/transform-index-html.test.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/basic-exports.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/basic-imports.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/comments.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/dynamic imports.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/file-extension.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/folder-index-js.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/import-in-code.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/import-meta.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/module-field.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/non-bare.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/preserve-symlink.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/resolve-symlink.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/urls.js
… packages/es-dev-server/test/snapshots/transform-index-html/import-map-all.html
… packages/es-dev-server/test/snapshots/transform-index-html/import-map-modern.html