https://github.com/open-wc/open-wc/pull/704 was merged, but I had to revert as I didn't cover some scenario's properly. That's fixed now in this PR.

The current implementation of node-resolve uses babel. This is wasteful, as we're creating a full AST of each js file just for resolving the imports.

In this implementation we use https://github.com/guybedford/es-module-lexer. This just extracts information about where imports and dynamic imports are, so that we can resolve them and do a simple string replacement.

Depending on the project, this can significantly speed up performance. Running the dev server and importing all of lodash-es takes 5200ms to load on master, and it takes 1800ms in this branch. Running large bundled libraries like sinon, chai, axe etc. through babel take much longer so that's a huge boost there.

On my machine, running the open-wc tests on master takes 5560ms while on this branch it takes 3670ms. This is without coverage turned on, because that still boots up babel. I also updated karma-esm to only run babel on files that need coverage. That brings it down to 4400ms with coverage enabled.

Fixes #664
Fixes https://github.com/open-wc/open-wc/issues/739